+++ src/main/kotlin/state/action/TraditionalActionBonus.kt
@@ -1,26 +1,26 @@
 package state.action
 
-import resource.RawResource
+import resource.raw.RawResource
 import state.State
 import kotlin.math.max
 
 data class TraditionalActionBonus(
     val type: ActionType,
-    val earnBase: Float,
-    val earnPerLevel: Float,
-    val earnFinal: Float,
-    val spentHunger: Float,
-    val spentThirst: Float,
-    val spentFeeling: Float,
+    val earnBase: Double,
+    val earnPerLevel: Double,
+    val earnFinal: Double,
+    val spentHunger: Double,
+    val spentThirst: Double,
+    val spentFeeling: Double,
 ) : ActionBonus {
     private val isEarningExp = type == ActionType.STUDY
 
     override fun getBonusDelta(status: State): ActionBonusDelta {
-        val efficiency = (if (status.hunger <= 25) 0.25f else 0.5f) + (if (status.thirst <= 25) 0.25f else 0.5f)
-        val earning = max(earnBase * efficiency + status.level * earnPerLevel * (efficiency - 0.5) * 2.0, 0.0).toFloat()
+        val efficiency = (if (status.hunger <= 25) 0.25 else 0.5) + (if (status.thirst <= 25) 0.25 else 0.5)
+        val earning = max(earnBase * efficiency + status.level * earnPerLevel * (efficiency - 0.5) * 2.0, 0.0)
         return ActionBonusDelta(
-            earnedMoney = if (isEarningExp) 0.0f else earning,
-            earnedExp = if (isEarningExp) earning else 0.0f,
+            earnedMoney = if (isEarningExp) 0.0 else earning,
+            earnedExp = if (isEarningExp) earning else 0.0,
             spentHunger = spentHunger,
             spentThirst = spentThirst,
             spentFeeling = spentFeeling
@@ -35,12 +35,12 @@ data class TraditionalActionBonus(
         fun fromRawResource(type: ActionType, rawResource: RawResource): TraditionalActionBonus {
             return TraditionalActionBonus(
                 type = type,
-                earnBase = rawResource["MoneyBase"]!!.toFloat(),
-                earnPerLevel = rawResource["MoneyLevel"]!!.toFloat(),
-                earnFinal = rawResource["FinishBonus"]!!.toFloat(),
-                spentHunger = rawResource["StrengthFood"]!!.toFloat(),
-                spentThirst = rawResource["StrengthDrink"]!!.toFloat(),
-                spentFeeling = rawResource["Feeling"]!!.toFloat(),
+                earnBase = rawResource["MoneyBase"]!!.toDouble(),
+                earnPerLevel = rawResource["MoneyLevel"]!!.toDouble(),
+                earnFinal = rawResource["FinishBonus"]!!.toDouble(),
+                spentHunger = rawResource["StrengthFood"]!!.toDouble(),
+                spentThirst = rawResource["StrengthDrink"]!!.toDouble(),
+                spentFeeling = rawResource["Feeling"]!!.toDouble(),
             )
         }
     }
